src/requirements.txt
pyyaml
pytest-flake8
requests
sqlalchemy
pandas
pyodbc
pyinstaller
sqlparse
flask-sqlalchemy